+++ back_joon/10951/boj1.testCase.js
@@ -1,0 +1,11 @@
+export const testCases1 = [
+    [1, 1],
+    [2, 3],
+    [3, 4],
+    [9, 8],
+    [5, 2],
+    [9, 9], // 조건 내 최대값
+    [-1, -5], // 음수
+    [100, 999], // 조건 초과
+    [1.11, 3.2], // 소수점
+];
