+++ back_joon/9093/boj2.testCase.js
@@ -1,15 +1,3 @@
-export const testCases1 = [
-    [1, 1],
-    [2, 3],
-    [3, 4],
-    [9, 8],
-    [5, 2],
-    [9, 9], // 조건 내 최대값
-    [-1, -5], // 음수
-    [100, 999], // 조건 초과
-    [1.11, 3.2], // 소수점
-];
-
 export const testCases2 = [
     '8',
     'I am happy today',
